app/requirements.txt
djangorestframework==3.13.1
Django==4.1.7
psycopg2-binary==2.9.5
django-mptt==0.13.4
django-mptt-admin==2.4.1